refactor: move Contentful Entry types to util file
