+++ src/lib/contentful/isDraft.ts
@@ -1,3 +1,3 @@
-import type { BlogPostEntry } from '@/components/Blog/Post'
+import type { BlogPostEntry } from '@/config/types'
 
 export const isDraft = (post: BlogPostEntry) => post.fields.isDraft
