+++ src/lib/contentful/isPressRelease.ts
@@ -1,6 +1,6 @@
 import { containsTag, PRESS_RELEASE_TAG } from '@/lib/containsTag'
 import { isDraft } from '@/lib/contentful/isDraft'
-import type { BlogPostEntry } from '@/components/Blog/Post'
+import type { BlogPostEntry } from '@/config/types'
 
 export const isPressReleasePost = (post: BlogPostEntry) => containsTag(post.fields.tags, PRESS_RELEASE_TAG)
